callbacks: s/PyObject_GetAttr+PyEval_CallObject/_PyObject_CallMethodIdObjArgs/
In order to take advantage of fast calls, stop looking up the attr,
generating an args list, and then invoking the method on the object.
Instead, generate a Py_Identifier and use that to invoke-by-id on the
object.  This assumes that any callbacks will exist, which we already
ensure, or else an AttributeError will be thrown.