callbacks: add dummy methods
This will simplify a future change to the callback code by ensuring that the
base object always has the right callback methods, even if they do nothing.
I do not know of a good way to disambiguate between an AttributeError due to
invoking a missing callback method and one thrown inside the callback.