execute: always re-raise exceptions
Relying on llhttp_execute to return HPE_USER when there is an exception is
not always reliable.  Instead, just check the error state and re-raise after
washing up, since anything that happened in a callback is more important
than llhttp's view of its own state and llhttp is able to complain again
later.