readme: update example re: header field names
HTTP header names are case insensitive and historically should be
interpreted as ISO-8859-1.  This change modifies the example to accumulate
them as bytearrays instead of bytes, then decodes them, lower-cases them,
and asserts that there is not a header name collision.  It also modifies
header values to be accumulated as bytearrays.